pyproject.toml
[tool.poetry]
name = "qstore"
version = "0.1.0"
description = ""
authors = ["Sino0on"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-cors-headers = "^4.5.0"
django-filter = "^24.3"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
pillow = "^11.0.0"
psycopg2-binary = "^2.9.10"
python-decouple = "^3.8"
jsonschema = "^4.23.0"
jsonschema-specifications = "^2024.10.1"
whitenoise = "^6.7.0"
django-debug-panel = "^0.8.3"
django-debug-toolbar = "^4.4.6"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
